Changed the capture equation and hole location
Previously, the hole location was set at the point where the ball
stopped rolling given the initial speed and direction. Now, it is set
at the point where the ball's velocity dips below 0.815 m/s, which is
the half the maximum velocity for a holed putt. In addition, the
puttWillBeMade function was modified so that putts with too much speed
will not be made.